+++ src/Delegate/NotFoundDelegate.php
@@ -8,10 +8,9 @@
 namespace Zend\Expressive\Delegate;
 
 use Fig\Http\Message\StatusCodeInterface;
-use Psr\Http\Message\RequestInterface;
+use Interop\Http\ServerMiddleware\DelegateInterface;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
-use Webimpress\HttpMiddlewareCompatibility\HandlerInterface as DelegateInterface;
 use Zend\Expressive\Template\TemplateRendererInterface;
 
 class NotFoundDelegate implements DelegateInterface
@@ -61,35 +60,12 @@ class NotFoundDelegate implements DelegateInterface
     }
 
     /**
-     * Proxy to handle method to support http-interop/http-middleware 0.1.1
-     *
-     * @param RequestInterface $request
-     * @return ResponseInterface
-     */
-    public function next(RequestInterface $request)
-    {
-        return $this->handle($request);
-    }
-
-    /**
-     * Proxy to handle method to support http-interop/http-middleware
-     * versions between 0.2 and 0.4.1
-     *
-     * @param ServerRequestInterface $request
-     * @return ResponseInterface
-     */
-    public function process(ServerRequestInterface $request)
-    {
-        return $this->handle($request);
-    }
-
-    /**
      * Creates and returns a 404 response.
      *
      * @param ServerRequestInterface $request
      * @return ResponseInterface
      */
-    public function handle(ServerRequestInterface $request)
+    public function process(ServerRequestInterface $request)
     {
         if (! $this->renderer) {
             return $this->generatePlainTextResponse($request);